third_party/WebKit/LayoutTests/external/wpt/longtask-timing/longtask-attributes-expected.txt

This is a testharness.js-based test.
FAIL Performance longtask entries are observable assert_equals: expected "same-origin-self" but got "self"
Harness: the test ran to completion.
